requirements/requirements_nn_train.txt
paho-mqtt==2.1.0
torch==2.4.0
scikit-learn==1.5.2
numpy==1.26.4
